More refactoring in partition/modularity analysis.
